Refactor level management into module
